Update many modules: bots, model, combat, cache
Massive set of changes across the codebase touching bots, model, combat, cache, launcher and server subsystems. Updates include extensive bot task/manager/tick logic and shop/trade tasks, numerous combat and special-attack definitions, cache/archive/store adjustments, launcher UI tweaks, plus assorted gameplay, NPC, and maintenance task updates to consolidate/refactor functionality.

+++ com/rs2/ConnectionThrottle.java
@@ -13,7 +13,7 @@ public final class ConnectionThrottle {
         if (!ConnectionThrottleSettings.connectionsEnabled) {
             return false;
         }
-        Integer n = connectionCountsByHost.putIfAbsent(string, 1);
+        Integer n = (Integer)connectionCountsByHost.putIfAbsent(string, 1);
         if (n != null && n == 3) {
             return false;
         }
